Finalizado os cadastros de Colaborador, Despesa e Filial

+++ exclusao/colaborador.php
@@ -1,13 +1,15 @@
 <?php
 
-require_once("../Model/colaborador.php");
-require_once("../config/Mensagens.php");
+require_once("Model/colaborador.php");
+require_once("config/Mensagens.php");
 
-$colaborador = new Colaborador();
+    $colaborador = new Colaborador();
+    $msg = new Mensagens();
 
-if(isset($_GET['param']))
+if(isset($url[2]))
 {
-    $id = intval(trim($_GET['param']));
+    $id = intval(trim($url[2]));
+    $colaborador->delete($id);
+    $msg::sucesso("Registro Excluido com Sucesso!","listagem/colaborador");
 }
 
-$colaborador->delete($id);
